unets-pytorch-lung/requirements.txt
pyqt5-sip==12.11.0
pyqt5==5.15.2
scipy==1.10.1
matplotlib==3.7.5
opencv_python==4.10.0.84
tqdm==4.67.1
Pillow==8.4.0
h5py== 3.11.0
numpy==1.23.1
setuptools==59.5.0
pillow==8.4.0
gradio==4.44.1
